Add new check to not delete events hwo already sended by email.

+++ files/app/Http/Controllers/Admin/Distribution.php
@@ -210,6 +210,11 @@ class Distribution extends Controller
                 continue;
             }
 
+            if ($distribution->isEmailSend) {
+                $canNotDelete++;
+                continue;
+            }
+
             $distribution->delete();
             $deleted++;
         }
